Multilayer perceptron predicts if have diabetes
This is a project I did for my Intro to Neural Networks class. For this project, I built a multilayer perceptron to try to predict whether a given individual has diabetes or not. The dataset we were supposed to use is an older, but commonly used  dataset,  the  Pima  Indians  Diabetes  Dataset.  Unfortunately, the highest accuracy my program was able to achieve was 85.71%. Things I changed to see if I could get a higher accuracy: number of hidden layers, number of nodes at each layer, different parameters in the compile() function (loss, metrics, loss_weights, etc), the optimizer, and the batch size.